requirements.txt
pyyaml==6.0.1
jsonschema==4.21.1